+++ src/Contacts/Contacts/Contact.java
@@ -27,5 +27,10 @@ public class Contact {
 
     public void setPhoneNum(Long phoneNum) {
         this.phoneNum = phoneNum;
+
     }
+
+//    public list showContacts(contactRoster){
+//
+//    }
 }
